+++ src/datasource-zabbix/zabbix/connectors/zabbix_api/zabbixAPIConnector.ts
@@ -4,10 +4,14 @@ import kbn from 'grafana/app/core/utils/kbn';
 import * as utils from '../../../utils';
 import { ZBX_ACK_ACTION_NONE, ZBX_ACK_ACTION_ADD_MESSAGE, MIN_SLA_INTERVAL } from '../../../constants';
 import { ShowProblemTypes, ZBXProblem } from '../../../types';
-import { GFHTTPRequest, JSONRPCError, ZBXScript, APIExecuteScriptResponse } from './types';
-import { getBackendSrv } from '@grafana/runtime';
+import { JSONRPCError, ZBXScript, APIExecuteScriptResponse } from './types';
+import { BackendSrvRequest, getBackendSrv } from '@grafana/runtime';
+import { rangeUtil } from '@grafana/data';
 
 const DEFAULT_ZABBIX_VERSION = '3.0.0';
+
+// Backward compatibility. Since Grafana 7.2 roundInterval() func was moved to @grafana/data package
+const roundInterval: (interval: number) => number = rangeUtil?.roundInterval || kbn.roundInterval || kbn.round_interval;
 
 /**
  * Zabbix API Wrapper.
@@ -42,18 +46,21 @@ export class ZabbixAPIConnector {
   //////////////////////////
 
   request(method: string, params?: any) {
-    return this.backendAPIRequest(method, params).then(response => {
-      return response?.data?.result;
-    });
+    if (!this.version) {
+      return this.initVersion().then(() => this.request(method, params));
+    }
+
+    return this.backendAPIRequest(method, params);
   }
 
-  backendAPIRequest(method: string, params: any = {}) {
-    const requestOptions: GFHTTPRequest = {
+  async backendAPIRequest(method: string, params: any = {}) {
+    const requestOptions: BackendSrvRequest = {
       url: this.backendAPIUrl,
       method: 'POST',
       headers: {
         'Content-Type': 'application/json'
       },
+      hideFromInspector: false,
       data: {
         datasourceId: this.datasourceId,
         method,
@@ -69,14 +76,15 @@ export class ZabbixAPIConnector {
       requestOptions.headers.Authorization = this.requestOptions.basicAuth;
     }
 
-    return getBackendSrv().datasourceRequest(requestOptions);
+    const response = await getBackendSrv().datasourceRequest(requestOptions);
+    return response?.data?.result;
   }
 
   /**
    * Get Zabbix API version
    */
   getVersion() {
-    return this.request('apiinfo.version');
+    return this.backendAPIRequest('apiinfo.version');
   }
 
   initVersion(): Promise<string> {
@@ -113,7 +121,7 @@ export class ZabbixAPIConnector {
       action: action
     };
 
-    if (severity) {
+    if (severity !== undefined) {
       params.severity = severity;
     }
 
@@ -142,7 +150,11 @@ export class ZabbixAPIConnector {
     return this.request('host.get', params);
   }
 
-  getApps(hostids): Promise<any[]> {
+  async getApps(hostids): Promise<any[]> {
+    if (semver.gte(this.version, '5.4.0')) {
+      return [];
+    }
+
     const params = {
       output: 'extend',
       hostids: hostids
@@ -161,11 +173,15 @@ export class ZabbixAPIConnector {
   getItems(hostids, appids, itemtype) {
     const params: any = {
       output: [
-        'name', 'key_',
+        'name',
+        'key_',
         'value_type',
         'hostid',
         'status',
-        'state'
+        'state',
+        'units',
+        'valuemapid',
+        'delay'
       ],
       sortfield: 'name',
       webitems: true,
@@ -195,11 +211,15 @@ export class ZabbixAPIConnector {
     const params = {
       itemids: itemids,
       output: [
-        'name', 'key_',
+        'name',
+        'key_',
         'value_type',
         'hostid',
         'status',
-        'state'
+        'state',
+        'units',
+        'valuemapid',
+        'delay'
       ],
       webitems: true,
       selectHosts: ['hostid', 'name']
@@ -651,6 +671,15 @@ export class ZabbixAPIConnector {
 
     return this.request('script.execute', params);
   }
+
+  getValueMappings() {
+    const params = {
+      output: 'extend',
+      selectMappings: "extend",
+    };
+
+    return this.request('valuemap.get', params);
+  }
 }
 
 function filterTriggersByAcknowledge(triggers, acknowledged) {
@@ -666,7 +695,7 @@ function filterTriggersByAcknowledge(triggers, acknowledged) {
 function getSLAInterval(intervalMs) {
   // Too many intervals may cause significant load on the database, so decrease number of resulting points
   const resolutionRatio = 100;
-  const interval = kbn.round_interval(intervalMs * resolutionRatio) / 1000;
+  const interval = roundInterval(intervalMs * resolutionRatio) / 1000;
   return Math.max(interval, MIN_SLA_INTERVAL);
 }