+++ src/datasource-zabbix/zabbix/zabbix.ts
@@ -1,5 +1,6 @@
 import _ from 'lodash';
 import moment from 'moment';
+import semver from 'semver';
 import * as utils from '../utils';
 import responseHandler from '../responseHandler';
 import { CachingProxy } from './proxy/cachingProxy';
@@ -21,18 +22,18 @@ const REQUESTS_TO_PROXYFY = [
   'getHistory', 'getTrend', 'getGroups', 'getHosts', 'getApps', 'getItems', 'getMacros', 'getItemsByIDs',
   'getEvents', 'getAlerts', 'getHostAlerts', 'getAcknowledges', 'getITService', 'getSLA', 'getVersion', 'getProxies',
   'getEventAlerts', 'getExtendedEventData', 'getProblems', 'getEventsHistory', 'getTriggersByIds', 'getScripts',
-  'getGlobalMacros'
+  'getGlobalMacros', 'getValueMappings'
 ];
 
 const REQUESTS_TO_CACHE = [
   'getGroups', 'getHosts', 'getApps', 'getItems', 'getMacros', 'getItemsByIDs', 'getITService', 'getProxies',
-  'getGlobalMacros'
+  'getGlobalMacros', 'getValueMappings'
 ];
 
 const REQUESTS_TO_BIND = [
   'getHistory', 'getTrend', 'getMacros', 'getItemsByIDs', 'getEvents', 'getAlerts', 'getHostAlerts',
-  'getAcknowledges', 'getITService', 'getVersion', 'acknowledgeEvent', 'getProxies', 'getEventAlerts',
-  'getExtendedEventData', 'getScripts', 'executeScript', 'getGlobalMacros'
+  'getAcknowledges', 'getITService', 'acknowledgeEvent', 'getProxies', 'getEventAlerts',
+  'getExtendedEventData', 'getScripts', 'executeScript', 'getGlobalMacros', 'getValueMappings'
 ];
 
 export class Zabbix implements ZabbixConnector {
@@ -42,6 +43,7 @@ export class Zabbix implements ZabbixConnector {
   getHistoryDB: any;
   dbConnector: any;
   getTrendsDB: any;
+  version: string;
 
   getHistory: (items, timeFrom, timeTill) => Promise<any>;
   getTrend: (items, timeFrom, timeTill) => Promise<any>;
@@ -56,8 +58,8 @@ export class Zabbix implements ZabbixConnector {
   getEventAlerts: (eventids) => Promise<any>;
   getExtendedEventData: (eventids) => Promise<any>;
   getMacros: (hostids: any[]) => Promise<any>;
-  getVersion: () => Promise<string>;
   getGlobalMacros: () => Promise<any>;
+  getValueMappings: () => Promise<any>;
 
   constructor(options) {
     const {
@@ -170,6 +172,17 @@ export class Zabbix implements ZabbixConnector {
     });
   }
 
+  async getVersion() {
+    if (!this.version) {
+      this.version = await this.zabbixAPI.initVersion();
+    }
+    return this.version;
+  }
+
+  supportsApplications() {
+    return this.version ? semver.lt(this.version, '5.4.0') : true;
+  }
+
   getItemsFromTarget(target, options) {
     const parts = ['group', 'host', 'application', 'item'];
     const filters = _.map(parts, p => target[p].filter);
@@ -220,7 +233,12 @@ export class Zabbix implements ZabbixConnector {
   /**
    * Get list of applications belonging to given groups and hosts.
    */
-  getAllApps(groupFilter, hostFilter) {
+  async getAllApps(groupFilter, hostFilter) {
+    await this.getVersion();
+    if (!this.supportsApplications()) {
+      return [];
+    }
+
     return this.getHosts(groupFilter, hostFilter)
     .then(hosts => {
       const hostids = _.map(hosts, 'hostid');
@@ -228,11 +246,14 @@ export class Zabbix implements ZabbixConnector {
     });
   }
 
-  getApps(groupFilter?, hostFilter?, appFilter?): Promise<AppsResponse> {
+  async getApps(groupFilter?, hostFilter?, appFilter?): Promise<AppsResponse> {
+    await this.getVersion();
+    const skipAppFilter = !this.supportsApplications();
+
     return this.getHosts(groupFilter, hostFilter)
     .then(hosts => {
       const hostids = _.map(hosts, 'hostid');
-      if (appFilter) {
+      if (appFilter && !skipAppFilter) {
         return this.zabbixAPI.getApps(hostids)
         .then(apps => filterByQuery(apps, appFilter));
       } else {