+++ src/test-setup/jest-setup.js
@@ -67,13 +67,6 @@ jest.mock('grafana/app/core/utils/datemath', () => {
   };
 }, {virtual: true});
 
-jest.mock('grafana/app/core/utils/kbn', () => {
-  return {
-    round_interval: n => n,
-    secondsToHms: n => n + 'ms'
-  };
-}, {virtual: true});
-
 jest.mock('grafana/app/core/table_model', () => {
   return class TableModel {
     constructor() {
@@ -99,6 +92,13 @@ jest.mock('grafana/app/core/config', () => {
 }, {virtual: true});
 
 jest.mock('jquery', () => 'module not found', {virtual: true});
+
+jest.mock('grafana/app/core/utils/kbn', () => {
+  return {
+    round_interval: n => n,
+    secondsToHms: n => n + 'ms'
+  };
+}, {virtual: true});
 
 // jest.mock('@grafana/ui', () => {
 //   return {};